rgw: move a few basic types into rgw_basic_types.{h,cc}
To help with the fight against circular dependencies
